Final merge from dev to master branch

+++ src/main/java/com/qa/ims/persistence/dao/Dao.java
@@ -4,17 +4,19 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 
+import com.qa.ims.persistence.domain.Order;
+
 public interface Dao<T> {
 
 	List<T> readAll();
 	
-	T read(Long id);
+	T read(int id);
 
 	T create(T t);
 
 	T update(T t);
 
-	int delete(long id);
+	int delete(int id);
 
 	T modelFromResultSet(ResultSet resultSet) throws SQLException;
 }
